Fix precision errors in frequency calculation

+++ src/engine/engine.h
@@ -320,10 +320,10 @@ class DivEngine {
     void setConf(String key, String value);
 
     // calculate base frequency/period
-    int calcBaseFreq(double clock, double divider, int note, bool period);
+    double calcBaseFreq(double clock, double divider, int note, bool period);
 
     // calculate frequency/period
-    int calcFreq(int base, int pitch, bool period=false, int octave=0);
+    int calcFreq(double base, double pitch, bool period=false, int octave=0);
 
     // find song loop position
     void walkSong(int& loopOrder, int& loopRow, int& loopEnd);
